requirements.txt
django==4.0
django-related-admin==1.0.1
black==21.12b0
